Move methods from Object to kernel
Specs fixed:

BasicObject subclass
- contains Kernel methods when including Kernel
