+++ spec/tags/core/basicobject/basicobject_tags.txt
@@ -1,4 +1,3 @@
 fails:BasicObject raises NameError when referencing built-in constants
 fails:BasicObject does not define built-in constants (according to defined?)
 fails:BasicObject includes itself in its list of constants
-fails:BasicObject subclass contains Kernel methods when including Kernel
