+++ topaz/modules/kernel.py
@@ -3,12 +3,13 @@ from __future__ import absolute_import
 import os
 import time
 
+from rpython.rlib.objectmodel import compute_identity_hash
 from rpython.rlib.rfloat import round_double
 from rpython.rlib.streamio import open_file_as_stream
 
 from topaz.coerce import Coerce
 from topaz.error import RubyError, error_for_oserror
-from topaz.module import ModuleDef
+from topaz.module import ModuleDef, check_frozen
 from topaz.modules.process import Process
 from topaz.objects.bindingobject import W_BindingObject
 from topaz.objects.exceptionobject import W_ExceptionObject
@@ -391,3 +392,77 @@ class Kernel(object):
         random_class = space.getclassfor(W_RandomObject)
         default = space.find_const(random_class, "DEFAULT")
         return default.srand(space, w_seed)
+
+    @moduledef.method("object_id")
+    def method_object_id(self, space):
+        return space.send(self, "__id__")
+
+    @moduledef.method("singleton_class")
+    def method_singleton_class(self, space):
+        return space.getsingletonclass(self)
+
+    @moduledef.method("extend")
+    def method_extend(self, space, w_mod):
+        if not space.is_kind_of(w_mod, space.w_module) or space.is_kind_of(w_mod, space.w_class):
+            if space.is_kind_of(w_mod, space.w_class):
+                name = "Class"
+            else:
+                name = space.obj_to_s(space.getclass(w_mod))
+            raise space.error(
+                space.w_TypeError,
+                "wrong argument type %s (expected Module)" % name
+            )
+        space.send(w_mod, "extend_object", [self])
+        space.send(w_mod, "extended", [self])
+
+    @moduledef.method("inspect")
+    def method_inspect(self, space):
+        return space.send(self, "to_s")
+
+    @moduledef.method("to_s")
+    def method_to_s(self, space):
+        return space.newstr_fromstr(space.any_to_s(self))
+
+    @moduledef.method("===")
+    def method_eqeqeq(self, space, w_other):
+        if self is w_other:
+            return space.w_true
+        return space.send(self, "==", [w_other])
+
+    @moduledef.method("send")
+    def method_send(self, space, args_w, block):
+        return space.send(self, "__send__", args_w, block)
+
+    @moduledef.method("nil?")
+    def method_nilp(self, space):
+        return space.w_false
+
+    @moduledef.method("hash")
+    def method_hash(self, space):
+        return space.newint(compute_identity_hash(self))
+
+    @moduledef.method("instance_variable_get", name="str")
+    def method_instance_variable_get(self, space, name):
+        return space.find_instance_var(self, name)
+
+    @moduledef.method("instance_variable_set", name="str")
+    @check_frozen()
+    def method_instance_variable_set(self, space, name, w_value):
+        space.set_instance_var(self, name, w_value)
+        return w_value
+
+    @moduledef.method("method")
+    def method_method(self, space, w_sym):
+        return space.send(
+            space.send(space.getclass(self), "instance_method", [w_sym]),
+            "bind",
+            [self]
+        )
+
+    @moduledef.method("tap")
+    def method_tap(self, space, block):
+        if block is not None:
+            space.invoke_block(block, [self])
+        else:
+            raise space.error(space.w_LocalJumpError, "no block given")
+        return self
